Prepare project for Backend TESTS
- Move route Controllers to separate directory to make testing possible
- Configure JEST library for testing
- Add 'mongodb-memory-server' library to provide testing MongoDB environment
