+++ package.json
@@ -33,6 +33,7 @@
   "devDependencies": {
     "concurrently": "^5.0.0",
     "jest": "^24.9.0",
+    "mongodb-memory-server": "^6.0.1",
     "nodemon": "^1.19.3"
   }
 }
